// src/app/models/employee.ts
export interface Employee {

  employeeId?:number;
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber: string;
  hireDate: string;
  jobTitle: string;
  salary: number;
  managerName: string;
  departmentName: string;
  country: string;
  city: string;
  street: string;
  postalCode: string;

}